+++ packages/atomic-react/src/components/EngineContext.ts
@@ -1,0 +1,8 @@
+import {SearchEngine} from '@coveo/headless';
+import {createContext} from 'react';
+
+export interface EngineContextType {
+  engine: SearchEngine;
+}
+
+export const EngineContext = createContext<Partial<EngineContextType>>({});
