+++ packages/atomic-react/src/components/index.ts
@@ -1,1 +1,4 @@
 export * from './stencil-generated/index';
+export {AtomicSearchInterfaceWrapper} from './SearchInterfaceWrapper';
+export * from '@coveo/headless';
+export {EngineContext, EngineContextType} from './EngineContext';
