+++ packages/samples/headless-react/src/pages/AtomicReactPage.tsx
@@ -1,9 +1,8 @@
 import {
   AtomicResultList,
   AtomicSearchBox,
-  AtomicSearchInterface,
+  AtomicSearchInterfaceWrapper,
 } from '@coveo/atomic-react';
-import {useEffect, useRef} from 'react';
 
 export function AtomicReactPage() {
   return (
@@ -12,33 +11,4 @@ export function AtomicReactPage() {
       <AtomicResultList />
     </AtomicSearchInterfaceWrapper>
   );
-}
-
-interface WrapperProps {
-  onReady?: () => void;
-}
-
-function AtomicSearchInterfaceWrapper(
-  props: React.PropsWithChildren<WrapperProps>
-) {
-  const searchInterface = useRef<HTMLAtomicSearchInterfaceElement>(null);
-
-  useEffect(() => {
-    init(searchInterface.current!);
-  }, []);
-
-  return (
-    <AtomicSearchInterface ref={searchInterface}>
-      {props.children}
-    </AtomicSearchInterface>
-  );
-}
-
-async function init(searchInterface: HTMLAtomicSearchInterfaceElement) {
-  await searchInterface.initialize({
-    accessToken: 'xx564559b1-0045-48e1-953c-3addd1ee4457',
-    organizationId: 'searchuisamples',
-  });
-
-  searchInterface.executeFirstSearch();
 }
